# 2x faster headers.get, headers.delete, headers.has