--- a/src/bun.js/bindings/webcore/HTTPHeaderMap.cpp
+++ b/src/bun.js/bindings/webcore/HTTPHeaderMap.cpp
@@ -219,6 +219,16 @@ bool HTTPHeaderMap::remove(const StringView name)
     if (findHTTPHeaderName(name, headerName))
         return remove(headerName);
 
+    return removeUncommonHeader(name);
+}
+
+bool HTTPHeaderMap::removeUncommonHeader(const StringView name)
+{
+#if ASSERT_ENABLED
+    HTTPHeaderName headerName;
+    ASSERT(!findHTTPHeaderName(name, headerName));
+#endif
+
     return m_uncommonHeaders.removeFirstMatching([&](auto& header) {
         return equalIgnoringASCIICase(header.key, name);
     });