### What does this PR do?

This makes headers.get(name), headers.delete(name) and headers.has(name) 2x faster in the happy path case where the header is well-known.

```js
bun on  jarred/faster-headers [$!?] via 🍞 v1.2.6
❯ bun bench/snippets/headers.mjs
clk: ~3.95 GHz
cpu: Apple M3 Max
runtime: bun 1.2.6 (arm64-darwin)

benchmark                     avg (min … max) p75   p99    (min … top 1%)
--------------------------------------------- -------------------------------
Headers.get('Content-Type')     16.35 ns/iter  16.30 ns        █
                        (14.94 ns … 21.38 ns)  18.88 ns ▁▁▁▁▁▁▅█▁▁▁▂▁▁▁▁▁▁▁▁▁
Headers.get('content-type')     16.72 ns/iter  16.68 ns       █
                        (15.37 ns … 27.70 ns)  19.52 ns ▁▁▁▁▁▄█▃▁▁▂▁▁▁▁▁▁▁▁▁▁
Headers.has('Content-Type')     14.87 ns/iter  14.83 ns      █
                        (13.43 ns … 26.38 ns)  17.76 ns ▁▁▁▁▂█▇▄▁▂▁▂▁▁▁▁▂▂▁▁▁
Headers.delete('Content-Type')  15.65 ns/iter  16.62 ns      ▄       █
                        (13.36 ns … 21.68 ns)  18.55 ns ▁▁▁▁▇██▃▃▄▄▃▄█▁▁▂▁▁▁▁

bun on  jarred/faster-headers [$!?] via 🍞 v1.2.6 took 5s
❯ bun-1.2.5 bench/snippets/headers.mjs
clk: ~3.89 GHz
cpu: Apple M3 Max
runtime: bun 1.2.5 (arm64-darwin)

benchmark                     avg (min … max) p75   p99    (min … top 1%)
--------------------------------------------- -------------------------------
Headers.get('Content-Type')     31.21 ns/iter  32.08 ns  ▃   ▄█▃▄
                        (28.45 ns … 43.65 ns)  37.03 ns █████████▆▄▅▅▅▃▂▂▂▂▁▁
Headers.get('content-type')     32.31 ns/iter  34.28 ns ▅▆▅██▆▄▃▂▃▄
                        (28.43 ns … 52.90 ns)  41.34 ns ████████████▆▄▃▂▂▁▁▁▁
Headers.has('Content-Type')     29.83 ns/iter  30.83 ns  ▇▆█▇█▆▄▂
                        (26.94 ns … 48.91 ns)  37.10 ns ▄█████████▇▄▄▃▂▂▂▁▂▁▂
Headers.delete('Content-Type')  30.63 ns/iter  31.38 ns ▂     █
                        (28.46 ns … 39.66 ns)  35.64 ns █▃▂▁▁▁█▃▂▂▁▁▇▂▂▁▁▁▁▁▁

bun on  jarred/faster-headers [$!?] via 🍞 v1.2.6 took 5s
❯ node bench/snippets/headers.mjs
clk: ~3.84 GHz
cpu: Apple M3 Max
runtime: node 23.9.0 (arm64-darwin)

benchmark                     avg (min … max) p75   p99    (min … top 1%)
--------------------------------------------- -------------------------------
Headers.get('Content-Type')     70.32 ns/iter  72.76 ns ▄█
                       (67.84 ns … 133.53 ns)  81.30 ns ██▇▄▂▂▂▂█▃▂▁▁▁▁▁▁▁▁▁▁
Headers.get('content-type')     57.89 ns/iter  59.88 ns █
                        (55.78 ns … 86.66 ns)  69.94 ns █▆▃▂▂▂█▃▂▂▁▁▁▁▁▁▁▁▁▁▁
Headers.has('Content-Type')     69.71 ns/iter  70.31 ns ▅▃ █▂
                       (66.59 ns … 136.12 ns)  82.80 ns █████▅▄▄▂▃▃▂▁▁▁▁▁▁▁▁▁
Headers.delete('Content-Type')  67.44 ns/iter  70.07 ns █
                       (65.26 ns … 136.61 ns)  77.87 ns █▆▂▃▁▂▂▂▆▃▂▁▁▁▁▁▁▁▁▁▁

bun on  jarred/faster-headers [$!?] via 🍞 v1.2.6 took 5s
❯ deno run -A bench/snippets/headers.mjs
clk: ~3.66 GHz
cpu: Apple M3 Max
runtime: deno 2.2.5+e2f66d4 (aarch64-apple-darwin)

benchmark                     avg (min … max) p75   p99    (min … top 1%)
--------------------------------------------- -------------------------------
Headers.get('Content-Type')    166.46 ns/iter 172.51 ns    █
                      (148.61 ns … 300.16 ns) 212.32 ns ▂▂▁██▄▃▃▃▃▃▂▂▁▂▁▁▁▁▁▁
Headers.get('content-type')    173.40 ns/iter 182.42 ns ▃▂  ▃▄██▇▆▂
                      (148.34 ns … 285.05 ns) 223.04 ns ██▆█████████▆▄▃▂▂▂▂▁▁
Headers.has('Content-Type')     54.91 ns/iter  55.53 ns       █
                       (49.07 ns … 218.96 ns)  69.15 ns ▂▁█▆▃▂█▅▂▂▂▂▁▁▁▁▁▁▁▁▁
Headers.delete('Content-Type') 180.44 ns/iter 208.00 ns           █
                       (83.00 ns … 182.63 µs) 250.00 ns ▁▁▁▁▁▁▁▁▁▁█▁▁▁▁▆▁▁▁▁▁
```

### How did you verify your code works?

Ran existing tests that seemed relevant. This should have no behavior difference.